raft: cleanup node (#65)
* remove optimisticUpdate


* remove MaxSizePerMsg


* rename


* remove node


* remove MsgUnreachable


* remove MustSync
