+++ kv/tikv/config/config.go
@@ -27,7 +27,7 @@ type Config struct {
 	RaftElectionTimeoutTicks    int
 	RaftMinElectionTimeoutTicks int
 	RaftMaxElectionTimeoutTicks int
-	RaftMaxSizePerMsg           uint64
+	RaftMaxEntsSize             uint64
 	RaftMaxInflightMsgs         int
 
 	// When the entry exceed the max size, reject to propose it.
@@ -125,7 +125,7 @@ func NewDefaultConfig() *Config {
 		RaftElectionTimeoutTicks:    10,
 		RaftMinElectionTimeoutTicks: 0,
 		RaftMaxElectionTimeoutTicks: 0,
-		RaftMaxSizePerMsg:           1 * MB,
+		RaftMaxEntsSize:             1 * MB,
 		RaftMaxInflightMsgs:         256,
 		RaftEntryMaxSize:            8 * MB,
 		RaftLogGCTickInterval:       10 * time.Second,
