+++ kv/tikv/raftstore/message/raft_router.go
@@ -11,5 +11,4 @@ type RaftRouter interface {
 	SendRaftMessage(msg *raft_serverpb.RaftMessage) error
 	SendRaftCommand(req *raft_cmdpb.RaftCmdRequest, cb *Callback) error
 	ReportSnapshotStatus(regionID uint64, toPeerID uint64, status raft.SnapshotStatus) error
-	ReportUnreachable(regionID, toPeerID uint64) error
 }
