+++ kv/tikv/raftstore/peer.go
@@ -140,12 +140,12 @@ func NewPeer(storeId uint64, cfg *config.Config, engines *engine_util.Engines, r
 		ID:            peer.GetId(),
 		ElectionTick:  cfg.RaftElectionTimeoutTicks,
 		HeartbeatTick: cfg.RaftHeartbeatTicks,
-		MaxSizePerMsg: cfg.RaftMaxSizePerMsg,
+		MaxEntsSize:   cfg.RaftMaxEntsSize,
 		Applied:       appliedIndex,
 		Storage:       ps,
 	}
 
-	raftGroup, err := raft.NewRawNode(raftCfg, nil)
+	raftGroup, err := raft.NewRawNode(raftCfg)
 	if err != nil {
 		return nil, err
 	}
