+++ raft/util_test.go
@@ -59,7 +59,6 @@ func TestIsLocalMsg(t *testing.T) {
 	}{
 		{pb.MessageType_MsgHup, true},
 		{pb.MessageType_MsgBeat, true},
-		{pb.MessageType_MsgUnreachable, true},
 		{pb.MessageType_MsgSnapStatus, true},
 		{pb.MessageType_MsgTransferLeader, false},
 		{pb.MessageType_MsgPropose, false},
